Add `canopy bulk --resume` option for resuming after error
Previously, if there was a small syntax error in a canopy bulk editing
session, in --sync mode they stuck around, but in regular editor mode
the edits were lost. This is a problem now especially given the new
`canopy review` feature that involves editing complex bulk files.

Now, the user is prompted to try `canopy bulk --resume` which takes
the existing bulk file and continues editing it.