Improve category topic pinning behavior with path substring removal
When rendering the bulk file, we alphabetize the topic files of a given
category. However, sometimes a given topic is a "hub" that indexes all
the other topics. Often, this topic has the same name as the category,
e.g. "topics/Literature/Hamlet/Hamlet.expl" which should even beat e.g.
"topic/Literature/Hamlet/Act_I_of_Hamlet.expl" because it is clearly
more definitionally related to the enclosing folder.

However, this relationship can be hard to capture via heuristic. E.g:

/topics/Literature/Shakespeare/Hamlet/Shakespeare's_Hamlet.expl

This should beat:

/topics/Literature/Shakespeare/Hamlet/Act_I_of_Hamlet.expl

Even though the latter is alphabetically prior. But, on what grounds?
Both contain the enclosing folder substring "Hamlet,", and the latter
might even be shorter.

So, we introduce a new heuristic, wherein we look at the enclosing
folder substrings, here "Literature," "Shakespeare," and "Hamlet",
and we remove them from the filenames when doing the