Error message: Fragment references cannot contain non-simple targets
Fragment references e.g. [#[Target]] allow a user to make a dummy local
reference with no child paragraph in order to allow other subtopics to
reference the fragment reference's specific point in its parent topic.

Previously, we didn't enforce that the target of a fragment link had to
be a simple target (e.g. "Target") and not a path (e.g. "Target/A#B/C),
which would be meaningless in this case, as the purpose of a fragment
reference is to create a subtopic in its parent topic, not make a real
reference to some path.

Now we error to alert the user of the problem.