Bug: HTML inclusion misindentified as alternate link syntax
Canopy supports the use of {{ABC}} syntax in two contexts:

1. In HTML, e.g. <b>{{[[New York]]}}</b>, double curly braces allow one
to evaluate the inner text as regular tokens such as links and insert
them under custom HTML.

2. In references, [[{{Dog}}s]], to allow identifying the link target as
some subset of the link's larger display text.

The bug is that you might have a reference that uses pipe syntax, and
then uses HTML with inclusions later, and these later curly braces
cause the link to be recognized as using {{ABC}} syntax as opposed to
pipe syntax, e.g.

[#[Second-Person Singular|אַתּ <br><br> _at_ <br><br><br> <small>{{You [man or woman\]}}</small>]]

We could attempt to disqualify curly braces inside HTML, but this
requires knowing where HTML ends, which might require full parsing, and
also HTML can include inclusions which contain references that contain
curly brace syntax, so you _can_ have valid reference curly brace
syntax even in HTML.

Perhaps it would be ideal to really parse references front to back,
adding characters to the various buffers and noting when a | or { or {{
is encountered not immediately inside HTML, although this might require
interaction with the HTML parsing code also to know when HTML is
finished or whether we are nested HTML > Insertion > Reference > Curly braces.

However, using contextual information about how Canopy is likely to be
used, we can find a simpler solution.

These are the patterns we are trying to identify:

[[A|<b>{{ABC}}]] // Pipe syntax with HTML inclusion

[[A{{B}}]] // Reference with curly brace syntax

[[{A|B}]] // Reference with ternary syntax, don't misidentify as pipe

We also know these facts:

1. We are very unlikely to find HTML insertions in the target text of
references, so, if we come across a {{ that is not in a pipe reference,
it is almost certainly HTML insertion.

2. The ternary syntax is not likely to be used to next complex HTML,
and so if we use a simple regex to find {a|b}, it is probably a ternary,
(and not e.g. a pipe in a reference in an HTML insertion in a ternary.)

3. Any { that is part of the reference target in a pipe reference is
unlikely to be matched, i.e. [[The letter {|Display text with a }]] is
unlikely to happen, so we can assume a simple check for \{.*\|.*\}
disqualifies a pipe reference.

So, putting those together, this procedure identifies all necessary cases:

1. If reference has pipe and pipe is not in a {a|b} pattern, it is a
pipe reference, and so subsequent {{ characters are part of HTML
insertions, or references which are within HTML insertions, and so
should be ignored as reference special characters.

2. If the reference does not have a pipe that is not in an {a|b} pattern,
then any subsequent {{ characters are reference syntax, and not HTML
insertions.

This procedure should handle all cases we are likely to see in the real
world, without requiring recursive parsing etc.