Fork server so that caching allows response for changes only
Previously the --cache option was introduced to allow finishing the
build of changed files before going on to parse unchanged files.
(Unchanged expl files may still have behavior changes because changed
files can affect existing path references i.e. rendering them invalid
due to obsolete adjacencies between other topics, etc.)

However, because we were serving from the same process, the author
still wouldn't be able to refresh changes until the entire parse was
finished.

Now, we always fork the server process, so that it doesn't wait on
the second pass, and additionally now we write the HTML after the first
pass – a second write isn't necessary because either the edit was to the
default topic, and so it got inlined correctly, or something has changed
in the unedited files, which can only be a new error, etc.

Additionally we consolidate naming around skipInitialBuild which is
more clearly an exception than the ostensibly clearer affirmative
initialBuild key.