Transfer link handler properly for menu cell links
If a menu cell contains a single link, we transfer the link handler
to the menu cell and not just the contained link. However, we had been
allowing link singletons to be children of other elements like <b>
but weren't then looking down the DOM for the listener on the descendant
link but rather the non-link direct child.