Fix issues emerging from shadow DOM feature
This commit fixes several issues that emerged from the shadow DOM
feature.

Canopy allows the user to request various paths, some of which
overlap with others in the work they have to add to the page, and so
initially we had cached this information simply in the DOM, having each
request check what existed already and request/render only the addition.
The shadow DOM feature moved the caching of existing request data into
a cache on the Paragraph class.

However, several issues were introduced by this change.

First, it became possible to have multiple versions of a given paragraph
object or sectionElement due to inconsistent use of the cached objects.
For example, one could generate a new sectionElement A for a given DOM
node B that was not currently in the DOM, then append some child C to
it, then take a parent node A, accidentally retrieve a cached version of
B that differed from the first, append it to A, and lose the appended C
that existed on the first version of B. To address this concern, we now
standardize use of the cache across the various methods a person could
access the paragraph class from, e.g. the Paragraph constructor,
Paragraph.for, etc, and ban instantiations that differ from the cached
version.

Second, various renders needed information for styling that would only
be available once in the DOM, whereas render occurs before insertion,
and so a helper was written to poll the DOM and run a styling callback
upon insertion. However, with the shadow DOM feature we can no longer
assume that a section element once rendered will ever be added to the
DOM, and certainly not within a given time frame, and so those polls
became infinitely repetitive etc. Now, we allow renderers to register
"postDisplayCallbacks" which will run if and when the sectionElement
is being displayed, post-DOM-integration, but before visual display.

A summary of the final cache mechanics is that Paragraph's constructor
is responsible for returning a unique paragraph object using the cache
but not updating it, and all writes to the cache occur in the
fetchAndRenderPath code upon final rendering of a given DOM tree.