Permit redundant parent links within paragraph and pluralize generally
We've gone back and forth between supporting paragraphs which have
multiple parent links, either two local references subsuming the same
paragraph or multiple global references in a paragraph to the same topic
or to paths beginning with the same topic etc.

Now we are supporting both, and this commit introduces various changes
necessary to do so. We had had a lot of logic around breaking ties for
which link should be parent, despite these being non-essential now,
they are preserved in Paragraph#parentLinks in case we ever want to
reintroduce any of that behavior, or need to pick one link to be parent
for some functionality.