Support Case-Insensitive URLs
Previously URLs were case-sensitive because we used the URL to get the
mixed-case values we needed to match local references to JSON subtopics
during DOM tree render, and also to store paragraphs in the
paragraphs by path cache on Paragraph.

Now, we make JSON filenames use Topic#caps to allow case-insensitive
versions to still hit, and then on retrieval we build a canonical list
of topic and subtopic names in requestJson, and use it to power a new
Path.recapitalize getter which checks the correct version. Luckily
fetchAndRenderPath already was set up to append and cache the tree
top-down, which means that we can assume at append-time that all
necessary ancestors are available to find a canonical path for that
stack frame.