Change bulk file pin selection to prefer closer directories
We do not track the position of topic files in their directories but
rather just use alphabetization for bulk file rendering to decide
order.

However, sometimes there is a file that represents the category and
which subsumes all other category contents. In such a case,
alphabetization is a confusing outcome. However, determining exactly
when to say a file is a "category topic" is tricky.

We could say it is the one that has the same name as the category
directory, but what if no file has a complete overlap and several have
partial overlaps?

Previously we had looked at which file name had the fewer extraneous
characters after removing the substring match with the container,
but this missed the fact that the extraneous characters left over from
a match with directory N may be strings from a match with directory
N - 1, e.g. Texas/Toyota/Toyota_of_Texas.expl should not lose to
Texas/Toyota/Toyota_Products.expl because its overlap with Texas is
considered more extraneous characters than "Products."

So, now we look for the largest match of the containing folder, and if
that fails, we continue up and break ties with the containing folder.