Revise table cell normalization code
This commit cleans up the table cell size standardization code, and
introduces a new formula for deciding whether to normalize heights;
now, we are able to adjust both the baseline minimal difference before
normalization occurs and also to what degree smaller elements should
have a lower threshold for normalization, so that large percentage
differences that connote small absolute pixel values do not prevent
normalizing.

Given current constants, these are element sizes and % difference
below which we normalize.

50px  - 65% (33px)
75px  - 48% (36px)
150px - 32% (48px)
200px - 28% (55px)
300px - 23% (70px)
500px - 20% (100px)

Also in this commit is code to standardize similar rows of a table to
the same height (there is no need to standardize cell heights of a row
because this is the HTML standard) however this feature is currently
disabled.