Add canopy bulk --sync --all option to pull in new topics from FS
Currently if you start a bulk sync session with all the files in the
project, newly added topic files don't get added to the bulk file
because the code assumes the original selection was just incidentally
all the files, but now that there are new ones, we meant to only
work on the previous subset. Now with the all option, we can pass the
--all flag to indicate new topics on disk should be added to the bulk
file.