Consolidate tryAndWriteHtml helpers and let sync throw translated error
When introducing bulk file translation, we needed a way for watch to
give the original error to bulk sync so that we can translate and
rethrow. So, we break up the helpers so that writing to disk can be
done as a separate helper and tryAndWriteHtml takes an error handler
and the option to suppress error throw so that the callback can throw.