Table normalization should consider colspan as multiple cells
If a cell spans multiple rows, it shouldn't be considered one big cell
to be ineligible for normalization, but rather should consider whether
the content in the spanning cell would be eligible for per-unit
normalization, and if so, snaps to an N-unit size.

E.g. for table | Two columns | One | One |

"Two columns" should snap if its content is close enough to 2x the
content size of the one-column cells.