Use URL-safe characters plus hash for JSON file names
Previously we URL encoded JSON file names and double encoded requests
but this was brittle and depended on server behavior. Now, we take any
URL-safe characters and put them in the file name followed by a hash
based on the full string to avoid collisions.

We also now use First_Letter_Capitalized_Formatting for the json file
name, which is less startling than caps.