Do resizing and cycle detection work in preDisplayCallbacks
Previously, renderers would append elements to a temporary section
element and put it in the DOM to test sizes/styles. Give the
render-time indeterminacy of downCycle links made it desirable to do
some of this work preDisplay post-DOM, not at render-time, now we are
doing all similar work preDisplay to avoid having to append things
to temporary DOM.

Part of the contract for this new timing to work is that all links
must be appended to the element returned by the renderer so that we
can find e.g. parentLinks, linkTo etc, and if sizing callbacks will
move them around e.g. menu rows, that's fine.

Now we cache the last valid value of Paragraph.selection and
Path.rendered for these sizing callbacks to get values during display
when CSS classes are unreliable.